working version of voice realtime interaction except data not being stored
