+++ generic_chatbot/wait-for-db.sh
@@ -12,7 +12,7 @@ done
 
 >&2 echo "✅ MySQL is up - proceeding with migrations"
 
-''' 
+
 # Run Django migrations
 echo "Running makemigrations..."
 
@@ -44,7 +44,7 @@ else
 fi
 
 >&2 echo "✅ All setup tasks completed. Starting application..."
-'''
+
 
 # Execute the original command (e.g., runserver)
 exec "$@"
